personalDataRegister final fix

+++ src/functions/handlers/hears/handlerPersonalDataRegister.ts
@@ -7,7 +7,11 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
   const chatID = ctx.chat?.id;
   console.log(ctx.session.messageId, "ctx.session.messageId");
   const menuAdress = menus["adress"];
+  const menuAdressEmpty = menus["adressEmpty"];
+  const menuAdressChange = menus["adressChange"];
   const menuPhone = menus["phone"];
+  const menuPhoneEmpty = menus["phoneEmpty"];
+  const menuPhoneChange = menus["phoneChange"];
   const messageInput = ctx.message?.text ? ctx.message.text : "";
   const isWaitingForPhone = ctx.session.isWaitingForPhone;
   const isWaitingForAdress = ctx.session.isWaitingForAdress;
@@ -16,11 +20,21 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
   const isPhoneCorrect = /^\+?[0-9\s\-\(\)]{6,}$/;
   const isAdressCorrect = messageInput.length > 10;
   const keyboardAdress = await createInlineKeyboard(menuAdress.buttons);
+  const keyboardAdressEmpty = await createInlineKeyboard(
+    menuAdressChange.buttons
+  );
+  const keyboardAdressChange = await createInlineKeyboard(
+    menuAdressEmpty.buttons
+  );
   const keyboardPhone = await createInlineKeyboard(menuPhone.buttons);
+  const keyboardPhoneEmpty = await createInlineKeyboard(menuPhoneEmpty.buttons);
+  const keyboardPhoneChange = await createInlineKeyboard(
+    menuPhoneChange.buttons
+  );
 
   if (isWaitingForPhone || isWaitingForAdress) {
     if (isWaitingForPhone && !isWaitingForAdress) {
-      if (isPhoneCorrect.test(messageInput)) {
+      if (isPhoneCorrect.test(messageInput) && messageInput) {
         ctx.session.phone = messageInput;
         ctx.session.isWaitingForPhone = false;
         const message = await ctx.reply(
@@ -29,16 +43,24 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
             reply_markup: keyboardPhone,
           }
         );
-        await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        if (ctx.session.messageId) {
+          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        } else {
+          console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+        }
         ctx.session.messageId = message.message_id;
         console.log(ctx.session.messageId, "Вы установили телефон:");
       } else {
         const message = await ctx.reply(
-          `❗️Ошибка ввода номера телефона\n\nВведите корректный номер телефона\n\nНомер  можно вводить в любом формате, миним 6 цифр, например\n\n89123123123 | +79123123123 | +7-912-312-31-13`
+          `❗️Ошибка ввода номера телефона\n\nВведите корректный номер телефона\n\nНомер  можно вводить в любом формате, миним 6 цифр, например\n\n89123123123 | +79123123123 | +7-912-312-31-13`,
+          { reply_markup: keyboardPhoneEmpty }
         );
-
         console.log(ctx.session.messageId, "Ошибка номера телефона");
-        await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        if (ctx.session.messageId) {
+          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        } else {
+          console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+        }
         ctx.session.messageId = message.message_id;
       }
     } else {
@@ -51,14 +73,23 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
             reply_markup: keyboardAdress,
           }
         );
-        await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        if (ctx.session.messageId) {
+          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        } else {
+          console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+        }
         ctx.session.messageId = message.message_id;
         console.log(ctx.session.messageId, "Вы установили адрес:");
       } else {
         const message = await ctx.reply(
-          `❗️Ошибка ввода адреса доставки\n\nВведите корректный адрес доставки\n\nАдрес должен содержать:\n📍Город\n📍 Улицу\n📍Номер дома (номер офиса / квартиры / помещения)\n\nНапример: г.Калининград, ул.Брамса д.45, оф.1`
+          `❗️Ошибка ввода адреса доставки\n\nВведите корректный адрес доставки\n\nАдрес должен содержать:\n📍Город\n📍 Улицу\n📍Номер дома (номер офиса / квартиры / помещения)\n\nНапример: г.Калининград, ул.Брамса д.45, оф.1`,
+          { reply_markup: keyboardAdressEmpty }
         );
-        await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        if (ctx.session.messageId) {
+          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+        } else {
+          console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+        }
         ctx.session.messageId = message.message_id;
         console.log(ctx.session.messageId, "Ошибка ввода адреса доставки");
       }
@@ -66,21 +97,37 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
   } else {
     if (isWaitingForPhoneChange || isWaitingForAdressChange) {
       if (isWaitingForPhoneChange && !isWaitingForAdressChange) {
-        if (isPhoneCorrect && messageInput) {
+        if (isPhoneCorrect.test(messageInput) && messageInput) {
           ctx.session.phone = ctx.message?.text;
           ctx.session.isWaitingForPhoneChange = false;
           const message = await ctx.reply(
             `Вы изменили телефон на: ${ctx.session.phone}`,
             {
-              reply_markup: keyboardPhone,
+              reply_markup: keyboardPhoneChange,
             }
           );
-          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+          if (ctx.session.messageId) {
+            await ctx.api.deleteMessage(
+              chatID as number,
+              ctx.session.messageId
+            );
+          } else {
+            console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+          }
           ctx.session.messageId = message.message_id;
         } else {
-          const message = await ctx.reply(`Введите корректный номер телефона`);
+          const message = await ctx.reply(`Введите корректный номер телефона`, {
+            reply_markup: keyboardPhoneChange,
+          });
+          if (ctx.session.messageId) {
+            await ctx.api.deleteMessage(
+              chatID as number,
+              ctx.session.messageId
+            );
+          } else {
+            console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+          }
           ctx.session.messageId = message.message_id;
-          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
         }
       } else {
         if (isAdressCorrect && messageInput) {
@@ -89,16 +136,31 @@ export async function handlerPersonalDataRegister(ctx: MyContext) {
           const message = await ctx.reply(
             `Вы изменили адрес на: ${ctx.session.adress}`,
             {
-              reply_markup: keyboardAdress,
+              reply_markup: keyboardAdressChange,
             }
           );
-          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+          if (ctx.session.messageId) {
+            await ctx.api.deleteMessage(
+              chatID as number,
+              ctx.session.messageId
+            );
+          } else {
+            console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+          }
           ctx.session.messageId = message.message_id;
         } else {
           const message = await ctx.reply(
-            `❗️Ошибка ввода адреса доставки\n\nВведите корректный адрес доставки\n\nАдрес должен содержать:\n📍Город\n📍 Улицу\n📍Номер дома (номер офиса / квартиры / помещения)\n\nНапример: г.Калининград, ул.Брамса д.45, оф.1`
+            `❗️Ошибка ввода адреса доставки\n\nВведите корректный адрес доставки\n\nАдрес должен содержать:\n📍Город\n📍 Улицу\n📍Номер дома (номер офиса / квартиры / помещения)\n\nНапример: г.Калининград, ул.Брамса д.45, оф.1`,
+            { reply_markup: keyboardAdressChange }
           );
-          await ctx.api.deleteMessage(chatID as number, ctx.session.messageId);
+          if (ctx.session.messageId) {
+            await ctx.api.deleteMessage(
+              chatID as number,
+              ctx.session.messageId
+            );
+          } else {
+            console.log("Ошибка удаления, ctx.session.messageId отсутствует");
+          }
           ctx.session.messageId = message.message_id;
         }
       }
